+++ src/app/core/sessions.rs
@@ -1,5 +1,6 @@
-use crate::app::SqlitePool;
+use crate::app::{models, SqlitePool};
 use eyre::Result;
+use rusqlite::params;
 use time::OffsetDateTime;
 
 #[derive(Clone)]
@@ -20,17 +21,40 @@ impl LiwanSessions {
         Ok(())
     }
 
-    /// Get the username associated with a session ID, if the session is still valid.
+    /// Get the user associated with a session ID, if the session is still valid.
     /// Returns None if the session is expired
-    pub fn get(&self, session_id: &str) -> Result<Option<String>> {
+    pub fn get(&self, session_id: &str) -> Result<Option<models::User>> {
         let conn = self.pool.get()?;
-        let mut stmt = conn.prepare_cached("select username, expires_at from sessions where id = ?")?;
-        let (username, expires_at): (String, OffsetDateTime) =
-            stmt.query_row([session_id], |row| Ok((row.get("username")?, row.get("expires_at")?)))?;
-        if expires_at < OffsetDateTime::now_utc() {
-            return Ok(None);
-        }
-        Ok(Some(username))
+
+        let mut stmt = conn.prepare_cached(
+            r#"--sql
+            select u.username, u.role, u.projects
+            from sessions s
+            join users u
+            on lower(u.username) = lower(s.username)
+            where
+                s.id = ?
+                and s.expires_at > ?
+        "#,
+        )?;
+
+        let user = stmt.query_row(params![session_id, time::OffsetDateTime::now_utc()], |row| {
+            Ok(models::User {
+                username: row.get("username")?,
+                role: row.get::<_, String>("role")?.try_into().unwrap_or_default(),
+                projects: row.get::<_, String>("projects")?.split(',').map(str::to_string).collect(),
+            })
+        });
+
+        user.map(Some).or_else(
+            |err| {
+                if err == rusqlite::Error::QueryReturnedNoRows {
+                    Ok(None)
+                } else {
+                    Err(err.into())
+                }
+            },
+        )
     }
 
     /// Delete a session